Created a placeholder angular 2 card container
I'll use this for testing the components in the bootstrapper. The purpose is just to provide something for them to inject.
