+++ source/components/cardContainer/pager/pager.ts
@@ -4,7 +4,7 @@ import * as angular from 'angular';
 import * as _ from 'lodash';
 
 import { IDataSource, IDataPager } from '../dataSources/index';
-import { CardContainerController } from '../cardContainer';
+import { CardContainerController } from '../cardContainer.ng1';
 
 export let moduleName: string = 'rl.ui.components.cardContainer.pager';
 export let componentName: string = 'rlPager';
